# src/spec/swagger.json
{"openapi":"3.0.0","paths":{"/health-check":{"get":{"operationId":"AppHealthCheck","parameters":[],"responses":{"200":{"description":""}}}},"/identity/config":{"get":{"operationId":"IdentityGetConfig","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigModel"}}}}},"tags":["identity"]}},"/price/latest/{tokenPair}":{"get":{"operationId":"PriceCoreGetLatest","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PriceModel"}}}}},"tags":["price"]}},"/price/latestKline/{tokenPair}/{interval}":{"get":{"operationId":"PriceCoreGetLatestKline","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}},{"name":"interval","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PriceKlineModel"}}}}},"tags":["price"]}},"/indicators/latest/{tokenPair}":{"get":{"operationId":"IndicatorsGetLatest","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenIndicatorsModel"}}}}},"tags":["indicators"]}},"/indicators/latest/{tokenPair}/{interval}":{"get":{"operationId":"IndicatorsGetLatestInterval","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}},{"name":"interval","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenIndicatorsModel"}}}}},"tags":["indicators"]}}},"info":{"title":"PriceEmitter","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"schemas":{"StringObject":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"ConfigModel":{"type":"object","properties":{"type":{"type":"string","default":"price"},"tokens":{"default":[],"type":"array","items":{"type":"string"}},"intervals":{"default":[],"type":"array","items":{"type":"string"}}},"required":["type","tokens","intervals"]},"PriceModel":{"type":"object","properties":{"tokenPair":{"type":"string"},"interval":{"type":"string"},"price":{"type":"string","default":"0"},"timestamp":{"type":"integer","default":0,"format":"int64"}},"required":["tokenPair","interval","price","timestamp"]},"PriceKlineModel":{"type":"object","properties":{"tokenPair":{"type":"string"},"interval":{"type":"string"},"price":{"type":"string","default":"0"},"timestamp":{"type":"integer","default":0,"format":"int64"},"timestamp_close":{"type":"integer","default":0,"format":"int64"},"price_open":{"type":"string","default":"0"},"price_low":{"type":"string","default":"0"},"price_high":{"type":"string","default":"0"},"volume":{"type":"string","default":"0"},"trades_number":{"type":"string","default":"0"},"asset_volume_quote":{"type":"string","default":"0"},"asset_volume_taker_base":{"type":"string","default":"0"},"asset_volume_taker_quote":{"type":"string","default":"0"}},"required":["tokenPair","interval","price","timestamp","timestamp_close","price_open","price_low","price_high","volume","trades_number","asset_volume_quote","asset_volume_taker_base","asset_volume_taker_quote"]},"TokenIndicatorsModel":{"type":"object","properties":{"tokenPair":{"type":"string"},"interval":{"type":"string"},"price":{"type":"string"},"timestamp":{"type":"integer","format":"int64"},"indicators":{"type":"object"}},"required":["tokenPair","interval","price","timestamp","indicators"]}}}}